object_db.go: replace decode with openDecode
In a subsequent commit, we will introduce a function to produce an
implementation of Object the concrete type of which is determined by the
type associated with the object named "SHA".

In order to do so, we would like to be able to call "o.decode()" without
having to open the object referred to above more than once (for example,
once to check the type, and the second time to read the object's
contents).

To make this possible, let's teach decode to accept an `*ObjectReader`
instead of an `sha []byte`, and defer the responsibility of calling
`o.open()` (which roughly translates a `[]byte` into an `*ObjectReader,
error)`), to an external caller.

To make calling `decode()` as easy as it was before this change, we'll
introduce `decodeOpen()`, which first calls `o.open()`, and then passes
on either (1) the resulting error, or (2) the result of calling
`o.decode()` with the newly opened object.