object_db.go: introduce Object()
Historically, callers have been able to access different objects stored
in the object database by (1) knowing the type of the object that they
desire to open and (2) calling the function associated with that type
(e.g., `Blob()`, `Commit()`, and so on) based on that information.

Sometimes, however, callers would like to open an object of known
existence but unknown type. Let's introduce `Object()` to do just that,
and construct a concrete implementation based on the type of object in
the header.