go.mod: initial commit
For a similar rationale to [1], let's build package
github.com/git-lfs/gitobj on Go 1.11, and use Go modules.

Since we were running builds on a variety of Go versions before
introducing modules, let's no longer do that and instead run on Go 1.11
only.

Doing so allows us to uniformly build outside of the $GOPATH, and avoids
having to maintain different checkout locations based on the version of
Go being used.

[1]: git-lfs/wildmatch@4bab7d7 (go.mod: initial commit, 2018-08-30)