storage: add a storage abstraction
There are several places where we want to expose an abstraction for
storage for the object database.  Add read-write and read-only storage
abstractions and a backend abstraction that encapsulates one or more of
these abstractions as interfaces that can be implemented.

Provide a method to determine whether the data read from the storage is
compressed, since pack data will be uncompressed for us, but most other
storage will not be.