errors: add an error for nonexistent objects
Depending on the storage backend used, we can return any of several
errors if an object doesn't exist.  For consistency, let's add an error
specifically for this condition and place it into a new errors package
so that it can be referred to from both the main package and the pack
package without import cycles.  Add a function in addition to quickly
determine if the error is due to a missing object.