backend: add standard backend types for filesystem and memory
Add two basic concrete backend types, one based on the filesystem using
fileStorer and pack.Storage, and the other based on memory using
memoryStorer.  This provides a standard set of backends that we can use
for reading and writing actual Git repositories or for testing using
memory locations.

Note that this code cannot live in the storage package, where one would
logically expect it to be, because the concrete types implementing it
are private to the package.