storer: change Open to return io.ReadCloser
In a future commit, we'll want to provide a storage interface that is
read-only.  If the Open method for storers returns a io.ReadWriteCloser,
then we'll be forced to implement a useless Write method.  Since we
don't actually need the ability to call Write, simply make the storer
interface return an io.ReadCloser instead, which lets us easily provide
the read-only interface we want.