object_db: use ReadCloser for uncompressed objects
When we open a new object using the object database, ensure that we use
a ReadCloser for the uncompressed object.  This ensures that we close
the file we've opened when reading loose objects so we don't run out of
file descriptors.

Note that we do use this code path for reading loose objects in some
cases because we hand it a ReadCloser which automatically decompresses
the object.