pack/set.go: close potentially-opened index files
In [1], we matched upstream Git's behavior more closely by (1) first
opening the corresponding *.idx, and _then_ opening the associated
*.pack, skipping over any pairs which had missing/corrupt index files.

In this commit, augment that behavior by closing a file descriptor we
may have opened, but is otherwise deemed by os.Open to be unusable for
one reason or another.

This path is unlikely in practice, but is a safe change nonetheless.

[1]: 561ed224 (pack/set: ignore packs without indices, 2018-11-29)