fixbug: find alternates may fails
`findAllBackends` will fails if current directory is not the repository
root path while the path in the alternates file is relative. And if
current directory is not right, latter `Scan` will lost the objects under
the alternates dbs. So change the relative alternate paths to `Abs`
paths to create the storage objects, then the latter `Scan` works well.

Signed-off-by: Sun Chao <sunchao9@huawei.com>