object_db: prevent FD leaks when writing blobs
When writing a blob, there are two places where we use temporary files.
First, we use them when we write the blob itself. Second, we use them
when we encode the buffer. In both of these cases, however, we failed to
close the file after we were done with it, leading to file descriptor
leaks until Go garbage-collected the items.

Add a helper function to close and remove these files and use it in each
of these places. Place the os.Remove statement after the Close
statement, since Windows will be happier if we don't try to remove a
file we have open.