pack: escape characters before using them in glob patterns
When we open a pack file, we use a glob pattern to find the pack files
that are in our pack directory and then attempt to open them.  This
works fine if our directory contains no characters that are special to
the shell, but works poorly when we handle characters that are special
in glob patterns, most notably the left square bracket.  In such a case,
our pattern doesn't actually match anything, so we don't open any pack
files.  Git LFS then complains about a missing object.

Escape all the characters which are special to the shell so that we can
handle these paths properly.  Use character classes instead of
backslashes because backslash escaping is not available on Windows.  For
the same reason, punt on handling backslashes on Unix systems because
there isn't a great way to handle them portably.