Add a CI workflow for GitHub Actions
Since we're no longer using Travis CI, let's switch to using GitHub
Action for CI.  Build on the same Go versions that Git LFS does.  Limit
us to Ubuntu only, since almost nothing in this library depends on
anything OS specific.  We can always add more OS versions in the future
if we determine that we see with compatibility problems.