object: make Decode take a hash.Hash as parameter
In most cases, we can determine the object ID algorithm in use from its
length without much difficulty.  However, for trees, the data is stored
in a binary format, and consequently it's not possible to correctly
parse a tree without knowing the hash algorithm in use.  Let's pass the
hash algorithm down through the Decode function so we can know the
correct length to parse.

While we're at it, increase the buffer length for the tree parsing code
and make it work for both SHA-1 and SHA-256.