object_writer: support different hashes
When we write an object, we'll need to know its hash in order to compute
the proper value.  Let's pass the caller the hash algorithm as well
as the writer to write to.

We don't need a generic set of options here because we're not likely to
gain additional options that affect our parsing of the object.  This is
the first format change in many years, so we can defer adding more
complex options until we need them.