backend: pass hash to the file system backend
In order to read packs and indices, we need to know the hash algorithm
in use, since the length of the hash will differ.  Pass the hash
algorithm to the file system backend instantiation function.  Since
we're moving to a new major version and can break compatibility, let's
also just turn the two constructors into one that takes all of the
required arguments.  If the user does not want alternates, they can pass
an empty string and we'll do the right thing.