tree: don't use system stat constants
On most Unix systems, the stat constants (S_IFMT and friends) that Git
uses are the same: that is, the system values are the same ones that Git
uses.  However, on zOS, this isn't the case, and the system constants
are different.

Let's add a set of our constants that are used by Git to determine the
type of an object that will be right on all platforms instead of using
the system constants.  Note that the tests already cover this case;
they'll just now pass on additional systems.