tree: use filenames only as tree entry keys
The Merge() function should only use filenames as the keys
of a tree object's entries, not filenames combined with file
modes, which at present can result in duplicate file entries
reported by "git fsck" as errors.

We therefore change to use only the entry's names as keys
when merging new and existing tree entries together, and
adjust the basic tree merge test such that it would fail
without this change.