tree: add IsLink method for TreeEntry class
Some callers may need to disambiguate between regular blob
objects and those which represent symbolic links, so we add
an IsLink() method to the TreeEntry class that returns true
when the object is a blob and a symlink, and false otherwise.

We also expand the test suite slightly to exercise this new
method.