commit: improve error messages
Right now, the error message when trying to parse an invalid tree or
parent header is hard to pin down because it's simply a generic hex
string parsing error.  Make the error easier to discover in the future
by adding some descriptive text to indicate to the user what the problem
is.