commit: fix parsing of mergetag headers
Git contains some multiline headers, including `gpgsig`,
`gpgsig-sha256`, and `mergetag` headers.  These headers start with a
normal line of data and then are continued after a newline by
indentation with a single space, which is stripped when computing the
body.

Our header parsing for commits could misparse these because we use
`strings.Fields`, which strips leading space.  As a result, if the
mergetag message contains a line starting with "tree " and then a
non-hex character, we'd fail parsing since we'd attempt to parse it as a
tree header.  Let's fix this by using `strings.Split` instead, which
doesn't strip space and therefore won't try to misparse this.

The fake commit is synthesized from
be122abe4bcd6d39b37892daae28c8bf5e4030fc in the Linux kernel repository.
Identifying information and most of the text are removed for privacy and
license reasons.